+++ visidata/loaders/fixed_width.py
@@ -46,8 +46,8 @@ class FixedWidthColumnsSheet(Sheet):
         # compute fixed width columns
         for i, j in columnize(list(r[0] for r in self.rows[:options.fixed_rows])):
             c = FixedWidthColumn('', i, j)
-            c.name = '_'.join(c.getValue(r) for r in self.rows[:options.headerlines])
+            c.name = '_'.join(c.getValue(r) for r in self.rows[:options.header])
             self.addColumn(c)
 
         # discard header rows
-        self.rows = self.rows[options.headerlines:]
+        self.rows = self.rows[options.header:]